FogBugz #98277: Enable CMD_BOOTZ in uboot
Enable the capability in uboot to load a standard Linux zImage, instead
of having going through the extra step of making a uImage.

Starting with Linux kernel v3.7, socfpga has enabled AUTO_ZRELADDR
that allows Linux to automatically determine the entry address of the
kernel image. However, this feature only works with the zImage
not uImage.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>