FogBugz #143936: Preloader failure when loading large image from SDMMC
Preloader fail to load large bootloader (discovered when loading
VxWorks bootloader which is 354kB). This is because the new SDMMC
driver is using DMA to speed up the transfer. It will populate larger
structure for DMA descriptor (for bigger file) and overflow the stack.
This patch will ensure smaller block count per transfer. It will
cause multiple transfer but avoiding stack overflow.

Signed-off-by: Chin Liang See <clsee@altera.com>