FogBugz #150299: Enable all devices within CV and AV Preloader
To enable all devices that connected to HPS and available on
existing Cyclone V and Arria V dev kit. This is to enable the OSes
that require access to these devices (such as SPIM0 and UART1).
These changes are only applicable to static handoff file only.
No changes to SOCEDS as it will always generate based on Qsys design.

Signed-off-by: Chin Liang See <clsee@altera.com>
---
Changes for v2
- Updated the commit message