FogBugz #158537: Updating U-Boot FPGA2SDRAM driver
To update the U-Boot FPGA2SDRAM enablement driver where applycfg
bit need to be set after write to fpgaportrst. At same time, the
applycfg bit can be set only the SDRAM is in idle mode. With
that, the code to set the applycfg will be execute in OCRAM

Signed-off-by: Chin Liang See <clsee@altera.com>
---
Changes for v2
- used macro for mask
- updated commit message