FogBugz #166933: Fixed OCRAM false double bit error
Ensure the stack pointer memory region is initialized during the
early run of the Preloader and U-Boot. This is to avoid the false
double bit error which might occurred during stack access. It is
due to the OCRAM controller always read in 64 bits manner. This issue
is spotted when we switch to new Altera version GNU toolchain.

Signed-off-by: Chin Liang See <clsee@altera.com>