FogBugz #173924: Adding SDRAM ECC Scrubbing using DMA
User can optional enable the Preloader to scrub the SDRAM memory
to initialize the SDRAM ECC bits. This is to enable subsequent
software that running on SDRAM enable the SDRAM ECC interrupt.

To do the scrubbing, DMA is used to speed up the scrubbing speed.
At same time, the DMA run in parallel while Preloader loading
the subsequent boot image from flashes.

To scrub the whole 1GB memory, it took 1446ms. As comparison,
scrubbing using processor and memset function is taking 13s.

Signed-off-by: Chin Liang See <clsee@altera.com>