Fogbugz #229396 Fix preloader handling of BSEL 0
When BootSelect is set to zero. The BootRom does almost nothing.
The BootRom does not setup any of the handoff to the preloader,
nor does the BootROM perform any work necessary to handle a warm reset.
This patch fixes the preloader's handling of BootSelect being zero.
When BootSelect is zero, the preloader will not examine any handoff
from the BootRom, and the preloader will execute a full cold reset
startup sequence.

Signed-off-by: Matthew Gerlach <mgerlach@opensource.altera.com>