FogBugz #227886: Setup SDRAM Protection Rule to prevent access beyond max mem
This patch enhances the previously submitted patch for FogBugz #223741 by
programming the SDRAM protection rules to prevent SDRAM accesses beyond
maximum configured SDRAM. This patch also corrects an issue in the way
SDRAM rows was calculated and configured in that same patch. The number
of rows is set such that 4G of memory is decoded to workaround the
SDRAM errata and limits memory accesses by using the SDRAM range protection
logic to cause a data abort if an access occurs beyond available,
configured memory in the system.

Signed-off-by: Vince Bridgers <vbridger@opensource.altera.com>