FogBugz #235540: Add CONFIG_MACH_TYPE for Cyclone 5 and Arria V
socfpga_cyclone5.h and socfpga_arria5.h are missing machine type definitions
- CONFIG_MACH_TYPE - for the Cyclone 5 and Arria 5 machine types. This is used
when buildling SPL/UBoot for Falcon mode, and is required. The Arria 5
machine ID was not registed on http://www.arm.linux.org.uk/developer/machines/
so part of the solution involved registering the Arria V machine type
with that website.

Signed-off-by: Vince Bridgers <vbridger@opensource.altera.com>