FogBugz #270904-24: fpga manager: more baby steps
This builds; fpga manager has not been tried on arria10 yet.

cfgwidth is hardwired to 32
cd_ratio is hardwired to x1

Note that we can't determine cfgwidth and cd_ratio from MSEL like we
did with cyclone5/arria5.  So for uboot we will probably need to add
uboot commands to set these (if not 32 and x1) before running commands
to load fpga.

Code contains //todo statements to point out stuff that needs to be done.

Signed-off-by: Alan Tull <atull@opensource.altera.com>