FogBugz #270904-13: fix sd/mmc dedicated IO
This commit brings the SD/MMC driver enablement a bit closer.
There is now a correct SD_CLK signal during the card ID procedure.

The 1st change the dedicated IO is to correctly enable the SD_PWR_ENA.
The 2nd change matches the SD signal configuration that the bootrom uses.

With this commit, the SD/MMC driver still cannot able to finish initializing,
but there is a correct SD_CLK signal.

Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>