FogBugz #270904-15: Configure OCRAM and DDR to be non-secure
Add a stub function to initialize a10 security policies for all logic units.
The idea is to set the security policy to normal, nonsecure for now.

In order for SD/MMC and USB to work, we need to make OCRAM and SDRAM
less secure so that data transfers can be accomplished.

Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>
Signed-off-by: Vince Bridgers <vbridger@opensource.altera.com>