FogBugz #270904-30: fpga manager: read header of rbf from fat, fixes and cleanup
On Arria10, in order to initialize the fpga manager, need to
determine the dclk ratio.  On Arria10, this is dependent on
whether the rbf is compressed and/or encoded.  Read the rbf
header to determine this.

As part of this change, fpgamgr_program_init() now takes a buffer
of the rbf header as its parameter.

Also misc fixes for fpga-manager and cff
 * NCE is active low, and PR_REQUEST is active high
 * Note that DCLK is not needed, apparantly using internal osc.
 * After sending the rbf data, watch for either:
   * condone_pin == 1 (success)
   * nstatus_pin == 0 (fail)
 * remove ifdef'ed out code
 * other code cleanup

Signed-off-by: Alan Tull <atull@opensource.altera.com>