FogBugz #270904-69: ignore null characters in serial input
During autoboot, ignore null character(s) in serial input after
programming FPGA.

Null characters were appearing in the uart fifo.  u-boot was
responding as if someone had hit a key during the countdown,
keeping uboot from continuing boot.  This was happening for
the uart on shared i/o, not for the uart on dedicated i/o.

Signed-off-by: Graham Moore <grmoore@opensource.altera.com>