FogBugz #270904-85: allow early use of shared io pins
Fix double call to ddr_calibration_sequence()

Add support for external-fpga-config boolean in the chosen node
of the device tree.  When present, the code will wait forever for the
fpga to be programmed before exiting the function, s_init().  The intention
is to allow early usage of a uart that might be connected to shared ios.

With the sd-card code configuration, and no external-fpga-config property nor
cff_filename property in the chose node,  the code will ignore all fpga
programming and drop to the uboot prompt.  This is intended for use with a uart
connected to dedicated pins or semihosting.  This allows for manually executing
the "fpga loadfs" and "ddrcal" commands.

With the qspi code configuration, external-fpga-config in the chosen
node is honored, and s_init() will not be exited until the fpga
is programed by an external source.  If external-fpga-config is not present
in the chosen node, streaming the fpga image from qspi is attempted.

Signed-off-by: Matthew Gerlach <mgerlach@opensource.altera.com>