FogBugz #282554 - move noc properties out of sdram node to noc node
The firewall registers are part of the NOC and not part of the sdram
controller.  So a NOC node was created in the device tree for the
the firewall information.  As part of this change the entire
sdram controller node in the device has been removed and the code
associated with it.  All sdram configuration is retreived at run time from
the NIOS.