FogBugz #292094: Arria10: fdtimagesize too small for GHRD DTB
The A10 uboot variable, fdtimagesize, is used as the number of
bytes to copy out of qspiflash for the linux device tree.
The patch changes the value from 0x5000 to 0x7f00 which is
the maximum space between the load addresses from the
dtb and kernel, 0x100 and 0x8000, respectively.

Signed-off-by: Matthew Gerlach <mgerlach@opensource.altera.com>