FogBugz #290468: Uboot device tree needs to be generated
The goal was always to check in a device tree generated
by the bootloader generator into the uboot source tree.
This device tree also has the correct GPLv2 licence required
by u-boot and approved by corporate counsel.  The generator
does not add a copyright, but the customer, in this case
Altera, puts on the desired copyright.

Signed-off-by: Matthew Gerlach <mgerlach@opensource.altera.com>