FogBugz #314587: Update DDR Scheduler ReadLatency calculation
The ReadLatency value is supposed to be set to the latency
out to SDRAM and back. Current calculation is considering
only the SDRAM portion of the total latency, and therefore
it is very under valued.  This will result in the DDR Scheduler
issuing a few requests and then it will stall waiting for
these requests to return, therefore limiting the bandwidth.

Signed-off-by: Chin Liang See <clsee@altera.com>