FogBugz #342152: Fixed the incorrect method on checking all bridges in idle status
This implementation rectified the method of checking to make sure all
bridges are in idle status by polling both noc_idleack and noc_idlestatus
registers are set to 1. The implementation of former method was not making
sure all bridges are in idle status. Using XOR instead of AND logic because
this is cleanest solution and shortest in coding.

Signed-off-by: Tien Fong Chee <tfchee@altera.com>