FogBugz #358687: Disable NOC timeout in A10 SoC
Disable NOC timeout that was left enabled by the function
reset_assert_all_bridges. The timeout is needed within that
function to shorten amount of time for idleack polling.

After the initialization, NOC timeout needs to be disabled.
This is to cater the scenario of long response time needed
by slow peripherals that reside in FPGA fabric.

Signed-off-by: Chin Liang See <clsee@altera.com>