FogBugz #367318: fix incorrect timer frequency
The 5 second count down timer was actually taking 10
seconds. The frequency of the input clock to
the timer had changed, but the u-boot configuration
did not track the change.  The clock frequency
is now calculated at run time to automatically
track any future change in the clock frequency.

Signed-off-by: Matthew Gerlach <mgerlach@opensource.altera.com>
Changes for v2
- Fix formatting