FogBugz #371156: fix "fpga loadfs" command hang
Prevent the command, fpga loadfs, from hanging by moving the
input error checking to the beginning of the function,
socfpga_loadfs(), before resets get asserted and an attempt
to program the fpga is made.  Added a check to ensure
the fpga is in Early Release mode before attempting to
program the "core" rbf.

Signed-off-by: Matthew Gerlach <mgerlach@opensource.altera.com>