FogBugz #371156-2: fix "fpga loadfs" command hang
This patch fixes the second hang in the command, fpga loadfs, when
the uart is connected to shared io pins.  When the initial peripheral
rbf is loaded, only the shared io pins are configured, and peripherals
connected to shared io pins are released from reset.  When the core
rbf is loaded, peripherals connected to shared io pins are left alone.
After the core or the combined rbfs are loaded, the fpga pins are
configured, and peripherals connected to those pins are released from reset.

Signed-off-by: Matthew Gerlach <mgerlach@opensource.altera.com>