FogBugz #371156-2: Printing the messages from buffer once UART is ready
Redirect the print messages to memory buffer when the UART is not ready.
Those messages stored in the buffer would be printed out to console
after FPGA image is programmed and UART with shared IO configuration done.
This would help to solve the UART stuck in transmit polling while
UART with shared IO held in reset.

Signed-off-by: Tien Fong Chee <tfchee@altera.com>
---
Changes for v2
- Replaced the pinmux UARTs offset value with #define MACRO.