FogBugz #540144-1: Add DIRECT/INDIRECT bit support
When sending mailbox command with large payloads from clients with
addressable memories such as the HPS processor it can be more efficient
to leave the data in its original location and arrange for the SDM
(Secure Device Manager) to fetch it directly from that location.
This is known as an indirect command. This patch allow function caller
to enable indirect command by setting the I bit (bit 11) to 1 in the mailbox
command header.

Signed-off-by: Chee Hong Ang <chee.hong.ang@intel.com>