FogBugz #544314: Ensure SiP function service table always align to 8 bytes.
There is an issue with memory alignment in the SiP function
service table '._secure_svc_tbl_entries' which cause the kernel
unable to boot when build with older version (6.4.1 and older)
of toolchain. This patch make sure the SiP function service
table always align to 8 bytes when build with older versions
of toolchain.

Signed-off-by: Chee Hong Ang <chee.hong.ang@intel.com>