rsa: read out public_exponent value based on 32-bit alignment
Public_exponent is a 64-bit data, which is stored in the device tree
blob in a 32-bit alignment address. This causes a problem for ARM64 when
public_exponent is read out one shot as a 64-bit data from an unaligned
address. Hence, to solve this problem, public_exponent is read out as two
32-bit datas, and then concatenating them.

Signed-off-by: Ooi, Joyce <joyce.ooi@intel.com>