HSD #1409755457-6: arm: dts: Stratix10: Modify stratix10 socdk memory node
The stratix10 socdk ships with 4GB of memory.  Modify the
device tree to represent this.  Note that to access 4GB of
memory in Stratix 10, due to the IO space from 2GB to 4GB,
we use the fact that the DDR controller ignores upper address
bits outside of the configured DRAM's size.  This means that
, the 4GB DRAM is mapped to memory every 4GB.

For an 8GB memory, you can either live with the 2GB IO space,
and loose access to that memory from the processor, or use
the same trick:

Loose 2GB of memory:
        memory {
		#address-cells = <2>;
		#size-cells = <1>;

                device_type = "memory";
                /* 8GB */
		/* first 2GB */
                reg = <0 0x00000000 0x80000000>,
		/* last 4GB */
                      <1 0x00000000 0x00000000>;
                u-boot,dm-pre-reloc;
        };

or to map it all:
        memory {
		#address-cells = <2>;
		#size-cells = <1>;

                device_type = "memory";
                /* 8GB */
		/* first 2GB */
                reg = <0 0x00000000 0x80000000>,
		/* next 6GB */
                      <2 0x80000000 0x80000000>;
                u-boot,dm-pre-reloc;
        };

Signed-off-by: Dalon Westergreen <dalon.westergreen@intel.com>
Signed-off-by: Ley Foon Tan <ley.foon.tan@intel.com>
Signed-off-by: Tien Fong Chee <tien.fong.chee@intel.com>