HSD #1707031258-3: arm: socfpga: stratix10: extend console command for RSU support
The Intel Remote System Update (RSU) provides a way for users to update
the QSPI configuration bitstream of a Intel Stratix 10 SoC device with
significantly reduced risk of corrupting the bitstream storage and
bricking the system.

This patch extends console commands to exercises the RSU APIs at
Intel Stratix10 SoC. The commands use the term "slot" to refer to a
sub-partition which is intended to contain a production image, and the
term "priority" to refer to the fact that the images are loaded by
firmware in the defined order.

The extended commands are below, you can refer to the header file rsu.h
for the details of RSU APIs:
	slot_count
	slot_by_name
	slot_get_info
	slot_size
	slot_priority
	slot_erase
	slot_program_buf
	slot_program_buf_raw
	slot_verify_buf
	slot_verify_buf_raw
	slot_enable
	slot_disable
	slot_load
	slot_load_factory
	slot_rename
	status_log

Also this patch removes the duplicated definition of struct
socfpga_rsu_s10_status.

Signed-off-by: Richard Gong <richard.gong@intel.com>
---
v2: changed to address review comments
v3: sorted do_rsu() as u_boot_cmd()
    changed to handle size in hex value
v4: replaced long 'if else if' with array of struct which includes
    command strings and function pointers
    updated the commit messages