ARM: socfpga: Synchronize the configuration for A10 NAND SoCDK
Update the default configuration file to enable the necessary
functionality the get the NAND SoCDK working. That includes fitImage
loading in SPL, booting both U-Boot and Linux.

The bridges need to be enabled for accessing the FPGA, so
enabling the bridges in A10 NAND default boot when FPGA is
in user mode.

Signed-off-by: Tien Fong Chee <tien.fong.chee@intel.com>