HSD #1707116690-1: arm: socfpga: Add fitImage for SPL programs periph RBF only
This is fitImage for programming peripheral RBF only in SPL, so that user
has option to program core RBF in U-Boot proper. The advantage of
programming core RBF in U-Boot proper especially for SDMMC boot the
performance can be increased up to around 90%, because program core RBF
in SPL hitting penalty on alignment issue in common FAT driver.

Signed-off-by: Tien Fong Chee <tien.fong.chee@intel.com>