HSD #1508140458: Fixed primary CPU core (CPU0) power on/off issue
This patch fix the issue of primary CPU core (CPU0) failed to come
online after being powered off (executed PSCI's 'CPU_OFF').

There are 2 arguments accepted by 'CPU_ON' which are 'cpu id' and
'entry_point'. When 'CPU_OFF' is called, it will clear 'cpu id' and
'entry_point' to zero. Since the 'cpu id' is set to 0, primary core
mistakenly thought it is being signaled to jump to the entry point
which happen to be NULL address.

This patch ensure 'entry_point' is non-NULL before it allows the
primary core to jump to the entry point to kick start the CPU again.

Signed-off-by: Chee Hong Ang <chee.hong.ang@intel.com>