HSD #14012726917: Replaced __udelay inline function with __socfpga_udelay inline function
The udelay function is used by both uboot proper and PSCI code
for timeout handling.The __udelay inline function was not included
in '__secure' section which causes random error when called by
PSCI code that run in '__secure' section.

The solution is replaced the __udelay inline function with
__socfpga_udelay inline function in timer.h, so that the code
can be included in '__secure' section.

Signed-off-by: Siew Chin Lim <elly.siew.chin.lim@intel.com>