configs: socfpga: soc64: Disable CONFIG_SPL_RAW_IMAGE_SUPPORT
Enable the SPL to support loading u-boot image with header
or in FIT format only. SPL shall proceed to the next boot
device in the boot order list when proper u-boot image is
unavailable in current device. This requires disabling the
raw image support in SPL.

This change is common for all Intel SOC64 devices.

Signed-off-by: Sin Hui Kho <sin.hui.kho@intel.com>