fpga: intel_pr: enable illegal request detection
This is to enable check on the freeze_illegal_request register
whereby high on any bit of this bus indicates a read or write
issue by a static regiion master when an Avalon-MM slave freeze
bridge is in the freeze state. In the case if any of the bit is
high, write 1 to first clear the bit and then return with error.

Signed-off-by: Chew, Chiau Ee <chiau.ee.chew@@intel.com>