HSD #15010294106: spl: fit: nand: fix fit loading on bad blocks
The offset at which the image to be loaded from NAND is located
is retrieved from the itb header. The presence of bad blocks in
the area of the NAND where the itb image is located could
invalidate the offset which must therefore be adjusted taking
into account the state of the sectors concerned.

Signed-off-by: Tien Fong Chee <tien.fong.chee@intel.com>