HSD 15011662064: Remove "x1" register to follow mainline changes
The mainline has removed the additional register "x1" in the
lowlevel_init_soc64.S due to the arm compiler upgrade. The downstream patch from
Siew Chin has the additional register. This causes the compilation to fail
during the u-boot-spl binary build. Since this is a temporary register, it is
safe to be removed from the branch_if_master in the Assembly code.

Signed-off-by: Lokanathan, Raaj <raaj.lokanathan@intel.com>