configs: defconfig: agilex_edge: Enable legacy boot flow build
This patch temporary enable the legacy boot flow to bypass missing
firewall issue on emulator. Enabled a copy of config for simics as
well to ease developer testing on Simics. Note that legacy boot flow
are not supported officially. Thus, this patch shall not merge into
any release branch. It shall be updated and removed in future once the
missing firewall issue resolved.

Signed-off-by: Sin Hui Kho <sin.hui.kho@intel.com>