configs: socfpga: sm: Temporary workaround to get SM legacy build passing
Legacy or Non atf flow is no longer supported in SM, so the new reset
driver added for SM support is only for ATF flow with SMCC to ATF for
both cold and warm reset request to SDM through mailbox executed in ATF.

This temporary workaround is for getting the legacy build passed because
this reset driver is only designed for ATF build flow. The
limitation of this change is cold reset with mailbox execution would
be supported in SPL and warm reset is not supported in legacy flow.

This patch should be removed when working on customer reference U-Boot
branch.

Signed-off-by: Tien Fong Chee <tien.fong.chee@intel.com>