drivers: mmc: sdhci-cadence: Fix the ComboPHY initialization
Fix: SDMMC does not come out of reset after the warm reboot in Simics.
The issue is due to the PHY udevice is empty and causing generic PHY_INIT
to fail. ComboPHY settings for assert/de-assert on SDMMC reset will be
skipped if the PHY_INIT function is not called.

Solution:
Initialize the PHY device first and assign the ComboPHY udevice to
the PHY struct. The PHY device will be checked in the generic PHY init
for proper PHY device bringup.

Signed-off-by: Kah Jing Lee <kah.jing.lee@intel.com>

---
v3:
- Rework based on review

v2:
- Initialize the ComboPHY using generic PHY function

v1:
- Rework, fix the ComboPHY init call
---