phy-cadence-combophy: Fix the ComboPHY initialization
Fix: SDMMC does not come out of reset after the warm reboot in Simics.
The issue is due to the PHY udevice is empty and causing generic PHY_INIT
to fail. ComboPHY settings for assert/de-assert on SDMMC reset will be
skipped if the PHY_INIT function is not called.

Solution: Fix the ComboPHY device init to retrieve the udevice from the
PHY device. The udevice is created during PHY device probe and stored
in the PHY struct.

Signed-off-by: Kah Jing Lee <kah.jing.lee@intel.com>