Revert "HSD #1507281912: spi: cadence_qspi: Probe fail if QSPI clock is not set"
This commit "39f5509b108533fc790746c07fa7c2a4fb2621ce" is reverted as this is
not required anymore. This was causing the compilation to fail.

With this changes, the cm_get_qspi_controller_clk_hz will decide the
refclock value and assigned it to the CONFIG_CQSPI_REF_CLK.

Signed-off-by: Lokanathan, Raaj <raaj.lokanathan@intel.com>