arch: mach-socfpga: cv_bsp_generator: fix support for Python3
Tested with generating a preloader for a DE10 nano board.

On newer systems (like Arch) getting Python2 back in-there is a bit of a
hassle (even if it is possible).
Also, on newer versions of Ubuntu, this is also going to be a bit more
difficult (over time).

Luckily, the conversion to Python3 is just a few lines.
A 'print' statement requires parentheses, and the 'unicode()' instantiation
can be converted to 'str()'. The 'unicode()' class is a derivative of
'str()', which is required in Python2 for unicode string handling; though,
it's unlikely we would need to generate unicode C definitions, so 'str()'
should be fine.

Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>