HSD #15014559963: image: Enable cache before reading DT
Enable cache before reading the device-tree.
This is a temporary workaround for SM A0 PO.

When FDT address is not specified, it will be randomly
picked by U-Boot and may be half-word aligned. When it
does, the half-word aligned FDT BE to LE swap will fail
as it uses LDR load register which requires word
alignment to load to memory when dcache is disabled.

Two ways to resolve this issue permanently -

1. Fix enable_caches - if the code for enable_caches is
temporary, we can set the load address of the FDT in the
.its file

2. Fix device tree - set the load address, must be word
aligned. Linux LZMA is loaded at offset 0x2000000, we
could try to load FDT at offset 0x3000000.

Signed-off-by: Alif Zakuan Yuslaimi <alif.zakuan.yuslaimi@intel.com>