configs: agilex: Add legacy defconfig
Add defconfig for legacy flow for Agilex. Legacy flow is the non-ATF
flow where u-boot.img is used instead of u-boot.itb which is a combination
of u-boot.img from U-Boot and bl31.bin from ATF

Keep this as internal for now as we may have some customers still using
this flow. To be re-visited later

Signed-off-by: Alif Zakuan Yuslaimi <alif.zakuan.yuslaimi@altera.com>