HSD #15018489098: spl: mmc: Select all raw-mode options
CycloneV is unable to boot from MMC starting v2025.01.

Comparing the .config between v2024.10 and v2025.01, it
was observed that CONFIG_SYS_MMCSD_RAW_MODE_U_BOOT_USE_PARTITION
and CONFIG_SYS_MMCSD_RAW_MODE_U_BOOT_USE_PARTITION_TYPE configs
are not set in v2025.10.

This is due to an external commit 2a00d73d081a ("spl: mmc: Try to clean
up raw-mode options") introduces a choice in the Kconfig where only one
of the three different ways to specify the offset is allowed.

In v2024.10, All three methods are enabled to specify the offset.

Follow the v2024.10 implementation to use all three methods
to specify the offset. This patch reverts the changes in the Kconfig
introduced by the mentioned commit above in order to get rid of the choice
in the Kconfig for SPL_SYS_MMCSD_RAW_MODE.

Signed-off-by: Alif Zakuan Yuslaimi <alif.zakuan.yuslaimi@altera.com>